renew_buf: unsigned int -> size_t

+++ Merger2to1A/Merger2to1A.cpp
@@ -302,7 +302,7 @@ int Merger2to1A::daq_run()
 }
 
 unsigned char * Merger2to1A::renew_buf(unsigned char * orig_buf,
-				      unsigned int cursize, unsigned int newsize)
+				      size_t cursize, size_t newsize)
 {
   unsigned char * new_buf;
