+++ Merger2to1B/Merger2to1B.cpp
@@ -302,7 +302,7 @@ int Merger2to1B::daq_run()
 }
 
 unsigned char * Merger2to1B::renew_buf(unsigned char * orig_buf,
-				      unsigned int cursize, unsigned int newsize)
+				      size_t cursize, size_t newsize)
 {
   unsigned char * new_buf;
